requirements.txt
# Variáveis de ambiente
python-dotenv

# Interface
gradio

# LangChain
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-google-genai

# Vetorização e PDFs
chromadb
pypdf
